PythonStudy/Smart2/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-filter==23.3
djangorestframework==3.14.0
Markdown==3.5.1
nose==1.3.7
pathlib==1.0.1
pytz==2023.3.post1
sqlparse==0.4.4
types-python-dateutil==2.8.19.14
tzdata==2023.3
